feat(useAPIForVADCDataDictionary): Created working prototype with endpoint connection and user authentication validation
